seperate out redux logic
